fsl/RMan: fix typecast compile warning
Fix typecast compile warning as below:

drivers/staging/fsl_rman/rman_uio_driver.c:89:9: note: in expansion of macro 'min'
  size = min(vma->vm_end - vma->vm_start, mem->size);
         ^

+++ drivers/staging/fsl_rman/rman_uio_driver.c
@@ -86,7 +86,7 @@ static int rman_uio_mmap(struct uio_info *info, struct vm_area_struct *vma)
 
 	mem = &info->mem[mi];
 
-	size = min(vma->vm_end - vma->vm_start, mem->size);
+	size = min(vma->vm_end - vma->vm_start, (unsigned long)mem->size);
 	size = max(size, PAGE_SIZE);
 
 	vma->vm_page_prot = pgprot_noncached(vma->vm_page_prot);
